Organize config + management command dirs
Rename directories to avoid confusing names like `src/src` and make management command consistent with actual implementation file (`cleanup_media` -> `cleanup_media.py`).

Follow-up of test suite reorganization.
