+++ .github/workflows/test-api.yml
@@ -15,7 +15,7 @@ on:
       - "src/tests/test_helpers.py"
       - "src/tests/fixtures/**"
       - "src/examples/**"
-      - "src/src/**"
+      - "src/config/**"
       - "src/manage.py"
       - "src/tool.jar"
       - "requirements.txt"
@@ -30,7 +30,7 @@ on:
       - "src/tests/test_helpers.py"
       - "src/tests/fixtures/**"
       - "src/examples/**"
-      - "src/src/**"
+      - "src/config/**"
       - "src/manage.py"
       - "src/tool.jar"
       - "requirements.txt"
