+++ src/app/management/__init__.py
@@ -1,4 +1,5 @@
 # SPDX-FileCopyrightText: 2026-present SPDX contributors
+# SPDX-FileType: SOURCE
 # SPDX-License-Identifier: Apache-2.0
 
 """Django management package for the app."""
