--- src/app/scripts/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: 2017-present SPDX contributors
-# SPDX-License-Identifier: Apache-2.0
-
-"""Django scripts package for the app."""
