+++ src/app/utils/cleanup_media.py
@@ -1,4 +1,5 @@
 # SPDX-FileCopyrightText: 2017-present SPDX contributors
+# SPDX-FileType: SOURCE
 # SPDX-License-Identifier: Apache-2.0
 
 """
