+++ src/config/.gitignore
@@ -1,3 +1,2 @@
-
 # environment files
 .env
