+++ src/config/version.py
@@ -1,4 +1,5 @@
 # SPDX-FileCopyrightText: 2020-present SPDX Contributors
+# SPDX-FileType: SOURCE
 # SPDX-License-Identifier: Apache-2.0
 
 """
